backend/requirements.txt
absl-py==2.3.1
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.69.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
asyncio==3.4.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
Bottleneck==1.6.0
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.12
fastapi-health==0.4.0
filelock==3.19.1
filetype==1.2.0
fireworks-ai==0.15.12
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.119.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.41.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphdatascience==1.15.1
greenlet==3.2.4
groq==0.32.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
httpx-ws==0.8.0
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-aws==0.2.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-fireworks==0.3.0
langchain-google-vertexai==2.0.25
langchain-groq==0.3.2
langchain-huggingface==0.3.0
langchain-neo4j==0.4.0
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langdetect==1.0.9
langserve==0.3.1
langsmith==0.3.45
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
multimethod==2.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.7.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
neo4j-rust-ext==5.28.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numexpr==2.13.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.23.0
openai==1.86.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.37.0
orjson==3.11.3
packaging==24.2
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
pip-install==1.3.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.1
pypandoc==1.15
pypandoc_binary==1.15
pyparsing==3.2.5
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
ragas==0.2.15
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Secweb==1.18.1
sentence-transformers==4.1.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
starlette-session==0.4.3
sympy==1.14.0
tenacity==9.1.2
textdistance==4.6.3
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.20
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.4.0
uvicorn==0.34.3
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.20.1
youtube-transcript-api==1.1.0
zipp==3.23.0
zstandard==0.23.0
